Semi-working build
There is an issue where the query is undefined after refresh... Lets see if this fixes it.
