+++ .vercel/output/functions/fn.func/.svelte-kit/output/server/chunks/internal.js
@@ -1,4 +1,4 @@
-import { c as create_ssr_component, a as setContext, v as validate_component, m as missing_component } from "./index.js";
+import { c as create_ssr_component, a as setContext, v as validate_component, m as missing_component } from "./index2.js";
 let base = "";
 let assets = base;
 const initial = { base, assets };
@@ -170,7 +170,7 @@ const options = {
 		<div class="error">
 			<span class="status">` + status + '</span>\n			<div class="message">\n				<h1>' + message + "</h1>\n			</div>\n		</div>\n	</body>\n</html>\n"
   },
-  version_hash: "1pe0393"
+  version_hash: "w824bf"
 };
 function get_hooks() {
   return {};
