+++ .vercel/output/functions/fn.func/.svelte-kit/output/server/entries/fallbacks/layout.svelte.js
@@ -1,4 +1,4 @@
-import { c as create_ssr_component } from "../../chunks/index.js";
+import { c as create_ssr_component } from "../../chunks/index2.js";
 const Layout = create_ssr_component(($$result, $$props, $$bindings, slots) => {
   return `${slots.default ? slots.default({}) : ``}`;
 });
