+++ .vercel/output/functions/fn.func/.svelte-kit/output/server/nodes/1.js
@@ -2,6 +2,6 @@
 
 export const index = 1;
 export const component = async () => (await import('../entries/fallbacks/error.svelte.js')).default;
-export const imports = ["_app/immutable/nodes/1.27a4f26c.js","_app/immutable/chunks/index.bd5c734a.js","_app/immutable/chunks/singletons.a2566be8.js"];
+export const imports = ["_app/immutable/nodes/1.321d7b81.js","_app/immutable/chunks/index.bd5c734a.js","_app/immutable/chunks/singletons.66258d8c.js"];
 export const stylesheets = [];
 export const fonts = [];
