+++ .vercel/output/functions/fn.func/.svelte-kit/vercel-tmp/manifest.js
@@ -4,7 +4,7 @@ export const manifest = {
 	assets: new Set([".nojekyll","favicon.png"]),
 	mimeTypes: {".png":"image/png"},
 	_: {
-		client: {"start":"_app/immutable/entry/start.01bad177.js","app":"_app/immutable/entry/app.e15af764.js","imports":["_app/immutable/entry/start.01bad177.js","_app/immutable/chunks/index.bd5c734a.js","_app/immutable/chunks/singletons.a2566be8.js","_app/immutable/entry/app.e15af764.js","_app/immutable/chunks/preload-helper.41c905a7.js","_app/immutable/chunks/index.bd5c734a.js"],"stylesheets":[],"fonts":[]},
+		client: {"start":"_app/immutable/entry/start.60f3b45e.js","app":"_app/immutable/entry/app.6486ec80.js","imports":["_app/immutable/entry/start.60f3b45e.js","_app/immutable/chunks/index.bd5c734a.js","_app/immutable/chunks/singletons.66258d8c.js","_app/immutable/entry/app.6486ec80.js","_app/immutable/chunks/preload-helper.41c905a7.js","_app/immutable/chunks/index.bd5c734a.js"],"stylesheets":[],"fonts":[]},
 		nodes: [
 			() => import('../output/server/nodes/0.js'),
 			() => import('../output/server/nodes/1.js')
