+++ .vercel/output/static/_app/immutable/nodes/2.92307466.js
@@ -1,0 +1,1 @@
+import{S as C,i as E,s as G,e as H,b as m,H as P,h as c,o as j,k as R,l as S,m as v,J as z,q as k,a as w,y as I,r as M,c as D,z as A,G as O,A as B,g as L,d as q,B as N}from"../chunks/index.bd5c734a.js";import{_ as Q}from"../chunks/preload-helper.41c905a7.js";const F=async({fetch:o})=>{const t=await(await o("/api/gridData")).json(),a=await(await o("/api/columnData")).json();return{grid:t,columns:a}},$=Object.freeze(Object.defineProperty({__proto__:null,load:F},Symbol.toStringTag,{value:"Module"}));function x(o){let n,t,r;return{c(){n=R("div")},l(a){n=S(a,"DIV",{}),v(n).forEach(c)},m(a,d){m(a,n,d),t||(r=z(o[1].call(null,n)),t=!0)},d(a){a&&c(n),t=!1,r()}}}function J(o){let n,t=o[0]&&x(o);return{c(){t&&t.c(),n=H()},l(r){t&&t.l(r),n=H()},m(r,a){t&&t.m(r,a),m(r,n,a)},p(r,[a]){r[0]?t||(t=x(r),t.c(),t.m(n.parentNode,n)):t&&(t.d(1),t=null)},i:P,o:P,d(r){t&&t.d(r),r&&c(n)}}}function K(o,n,t){let{gridData:r}=n,{columns:a}=n;console.log(r);let d=!1,u=!1,_=!1,s;j(()=>{(async()=>{const i=await Q(()=>import("../chunks/index.b25c9b93.js"),[],import.meta.url);t(6,s=i.default),t(5,u=!0),t(4,d=!0)})()});function g(i,e,l,f,y,b,T){s.renderers.TextRenderer.apply(this,arguments);let V=r[l].ToolQty;parseFloat(b,10)>V&&(e.style.background="red")}function p(i,e,l,f,y,b,T){s.renderers.TextRenderer.apply(this,arguments),!b||b===""||b==="Gigabox"&&(e.style.fontStyle="italic"),e.style.background=""}function h(i){console.log("Good!"),s.renderers.registerRenderer("negativeValueRenderer",p),s.renderers.registerRenderer("priceRenderer",g);try{new s(i,{data:r,columns:a,width:"100%",height:"auto",colHeaders:!0,rowHeaders:!1,colWidths:100,columnSorting:!0,filters:!0,dropdownMenu:!0,manualColumnResize:!0,licenseKey:"non-commercial-and-evaluation",afterSelection(e,l,f,y){this.getCellMeta(f,y).readOnly?this.updateSettings({fillHandle:!1}):this.updateSettings({fillHandle:!0})},cells(e,l){const f={},y=this.instance.getData();return(e===0||y[e]&&y[e][l]==="readOnly")&&(f.readOnly=!0),l>5&&(f.renderer="priceRenderer"),f}})}catch(e){console.log("Handsontable initialization error:",e)}}return o.$$set=i=>{"gridData"in i&&t(2,r=i.gridData),"columns"in i&&t(3,a=i.columns)},o.$$.update=()=>{o.$$.dirty&48&&t(0,_=d&&u),o.$$.dirty&64&&s&&console.log(s)},[_,h,r,a,d,u,s]}class W extends C{constructor(n){super(),E(this,n,K,J,G,{gridData:2,columns:3})}}function U(o){let n,t,r,a,d,u,_,s,g,p,h,i;return a=new W({props:{gridData:o[0],columns:o[1]}}),{c(){n=R("span"),t=k("Grid is here:"),r=w(),I(a.$$.fragment),d=w(),u=R("br"),_=w(),s=R("br"),g=w(),p=R("div"),h=k("production days that are producing over the tool capacity will be highlighted red")},l(e){n=S(e,"SPAN",{});var l=v(n);t=M(l,"Grid is here:"),l.forEach(c),r=D(e),A(a.$$.fragment,e),d=D(e),u=S(e,"BR",{}),_=D(e),s=S(e,"BR",{}),g=D(e),p=S(e,"DIV",{});var f=v(p);h=M(f,"production days that are producing over the tool capacity will be highlighted red"),f.forEach(c)},m(e,l){m(e,n,l),O(n,t),m(e,r,l),B(a,e,l),m(e,d,l),m(e,u,l),m(e,_,l),m(e,s,l),m(e,g,l),m(e,p,l),O(p,h),i=!0},p:P,i(e){i||(L(a.$$.fragment,e),i=!0)},o(e){q(a.$$.fragment,e),i=!1},d(e){e&&c(n),e&&c(r),N(a,e),e&&c(d),e&&c(u),e&&c(_),e&&c(s),e&&c(g),e&&c(p)}}}function X(o,n,t){let{data:r}=n;console.log(r.grid),console.log(r.columns);let a=r.grid,d=r.columns;return o.$$set=u=>{"data"in u&&t(2,r=u.data)},[a,d,r]}class ee extends C{constructor(n){super(),E(this,n,X,U,G,{data:2})}}export{ee as component,$ as universal};
