--- .vercel/output/static/_app/immutable/nodes/2.c3eded1f.js
@@ -1,1 +1,0 @@
-import{S as Q,i as A,s as E,e as P,b as f,H as V,h as c,o as G,k as h,l as S,m as w,J as M,q as x,a as v,y as O,r as C,c as L,z as I,G as H,A as T,g as U,d as z,B}from"../chunks/index.bd5c734a.js";import{_ as q}from"../chunks/preload-helper.41c905a7.js";function N(i){let r,t,a;return{c(){r=h("div")},l(n){r=S(n,"DIV",{}),w(r).forEach(c)},m(n,d){f(n,r,d),t||(a=M(i[1].call(null,r)),t=!0)},d(n){n&&c(r),t=!1,a()}}}function W(i){let r,t=i[0]&&N(i);return{c(){t&&t.c(),r=P()},l(a){t&&t.l(a),r=P()},m(a,n){t&&t.m(a,n),f(a,r,n)},p(a,[n]){a[0]?t||(t=N(a),t.c(),t.m(r.parentNode,r)):t&&(t.d(1),t=null)},i:V,o:V,d(a){t&&t.d(a),a&&c(r)}}}function F(i,r,t){let{data:a}=r,{columns:n}=r;console.log(a);let d=!1,m=!1,y=!1,s;G(()=>{(async()=>{const o=await q(()=>import("../chunks/index.b25c9b93.js"),[],import.meta.url);t(6,s=o.default),t(5,m=!0),t(4,d=!0)})()});function _(o,e,l,u,g,R,D){s.renderers.TextRenderer.apply(this,arguments);let k=a[l].toolQty;console.log(k),parseFloat(R,10)>k&&(e.style.background="red")}function p(o,e,l,u,g,R,D){s.renderers.TextRenderer.apply(this,arguments),!R||R===""||R==="Gigabox"&&(e.style.fontStyle="italic"),e.style.background=""}function b(o){console.log("Good!"),s.renderers.registerRenderer("negativeValueRenderer",p),s.renderers.registerRenderer("priceRenderer",_);try{new s(o,{data:a,columns:n,width:"100%",height:"auto",colHeaders:!0,rowHeaders:!1,colWidths:100,columnSorting:!0,filters:!0,dropdownMenu:!0,manualColumnResize:!0,licenseKey:"non-commercial-and-evaluation",afterSelection(e,l,u,g){this.getCellMeta(u,g).readOnly?this.updateSettings({fillHandle:!1}):this.updateSettings({fillHandle:!0})},cells(e,l){const u={},g=this.instance.getData();return(e===0||g[e]&&g[e][l]==="readOnly")&&(u.readOnly=!0),l>5&&(u.renderer="priceRenderer"),u}})}catch(e){console.log("Handsontable initialization error:",e)}}return i.$$set=o=>{"data"in o&&t(2,a=o.data),"columns"in o&&t(3,n=o.columns)},i.$$.update=()=>{i.$$.dirty&48&&t(0,y=d&&m),i.$$.dirty&64&&s&&console.log(s)},[y,b,a,n,d,m,s]}class J extends Q{constructor(r){super(),A(this,r,F,W,E,{data:2,columns:3})}}function K(i){let r,t,a,n,d,m,y,s,_,p,b,o;return n=new J({props:{data:i[0],columns:i[1]}}),{c(){r=h("span"),t=x("Grid is here:"),a=v(),O(n.$$.fragment),d=v(),m=h("br"),y=v(),s=h("br"),_=v(),p=h("div"),b=x("production days that are producing over the tool capacity will be highlighted red")},l(e){r=S(e,"SPAN",{});var l=w(r);t=C(l,"Grid is here:"),l.forEach(c),a=L(e),I(n.$$.fragment,e),d=L(e),m=S(e,"BR",{}),y=L(e),s=S(e,"BR",{}),_=L(e),p=S(e,"DIV",{});var u=w(p);b=C(u,"production days that are producing over the tool capacity will be highlighted red"),u.forEach(c)},m(e,l){f(e,r,l),H(r,t),f(e,a,l),T(n,e,l),f(e,d,l),f(e,m,l),f(e,y,l),f(e,s,l),f(e,_,l),f(e,p,l),H(p,b),o=!0},p:V,i(e){o||(U(n.$$.fragment,e),o=!0)},o(e){z(n.$$.fragment,e),o=!1},d(e){e&&c(r),e&&c(a),B(n,e),e&&c(d),e&&c(m),e&&c(y),e&&c(s),e&&c(_),e&&c(p)}}}function j(i){return[[{type:"Sales Order Demand",platform:"737",rowLabels:"213A3210-7",luLine:"LU280",toolQty:2,safetyStock:20,c06232022:2,c06242022:0,c06272022:3},{type:"Current MPS",platform:"737",rowLabels:"213A3210-7",luLine:"LU280",toolQty:2,safetyStock:20,c06232022:2,c06242022:2,c06272022:1},{type:"NEW MPS Suggestion",platform:"737",rowLabels:"213A3210-8",luLine:"LU280",toolQty:2,safetyStock:20,c06232022:2,c06242022:0,c06272022:3}],[{title:" ",type:"text",data:"type",renderer:"negativeValueRenderer"},{title:"Platform",type:"text",data:"platform",renderer:"negativeValueRenderer"},{title:"Row Labels",type:"text",data:"rowLabels",renderer:"negativeValueRenderer"},{title:"LU Line",type:"text",data:"luLine",renderer:"priceRenderer"},{title:"Tool Qty",type:"numeric",data:"toolQty",renderer:"negativeValueRenderer"},{title:"Safety Stock",type:"numeric",data:"safetyStock",renderer:"negativeValueRenderer"},{title:"06/23/2022",type:"numeric",data:"c06232022",className:"htCenter",renderer:"negativeValueRenderer"},{title:"06/24/2022",type:"numeric",data:"c06242022",className:"htCenter",renderer:"negativeValueRenderer"},{title:"06/27/2022",type:"numeric",data:"c06272022",className:"htCenter",renderer:"negativeValueRenderer"},{title:"etc",type:"numeric",data:"etc",className:"htCenter",renderer:"negativeValueRenderer"}]]}class Z extends Q{constructor(r){super(),A(this,r,j,K,E,{})}}export{Z as component};
