+++ src/routes/+page.js
@@ -1,8 +1,12 @@
 export const load = async ({ fetch }) => {
-    const response = await fetch(`/api/gridData`)
-    const grid = await response.json()
+    const responseMPS = await fetch(`/api/gridData`)
+    const grid = await responseMPS.json()
+
+    const responseCol = await fetch(`/api/columnData`)
+    const columns = await responseCol.json()
   
-      return {
-        grid
-      };
-    }
+    return {
+        grid,
+        columns
+    };
+}
